requirements.txt.txt
Django==4.2
playwright==1.39.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
requests==2.31.0
django-environ==0.11.2
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-dotenv==1.0.0
aiohttp>=3.9.0
